requirements.txt
amqp==2.4.2
aniso8601==6.0.0
astroid==2.2.5
Babel==2.6.0
billiard==3.6.0.0
celery==4.3.0
Click==7.0
colorama==0.4.1
cycler==0.10.0
dnspython==1.16.0
eventlet==0.25.0
Flask==1.0.3
Flask-RESTful==0.3.7
flower==0.9.3
greenlet==0.4.15
isort==4.3.20
itsdangerous==1.1.0
Jinja2==2.10.1
joblib==0.13.2
kiwisolver==1.1.0
kombu==4.5.0
lazy-object-proxy==1.4.1
MarkupSafe==1.1.1
matplotlib==3.1.0
mccabe==0.6.1
monotonic==1.5
mysql-connector-python==8.0.16
mysqlclient==1.4.2.post1
numpy==1.16.3
pandas==0.24.2
protobuf==3.7.1
pylint==2.3.1
PyMySQL==0.9.3
pyparsing==2.4.0
python-dateutil==2.8.0
python-dotenv==0.10.2
pytz==2019.1
scikit-learn==0.21.1
scipy==1.3.0
six==1.12.0
SQLAlchemy==1.3.3
tornado==5.1.1
typed-ast==1.3.5
vine==1.3.0
Werkzeug==0.15.4
wrapt==1.11.1
